requirements.txt
beautifulsoup4~=4.12.3
requests~=2.31.0
validators~=0.22.0
python-dotenv~=1.0.1
aiogram~=3.4.1
loguru~=0.7.2
pymongo~=4.6.2
dnspython~=2.6.1